refactor(shared): move audit entities to entities package and update imports
