+++ mark/src/main/java/pt/estga/mark/entities/MarkOccurrence.java
@@ -4,7 +4,7 @@ import jakarta.persistence.*;
 import lombok.*;
 import pt.estga.file.entities.MediaFile;
 import pt.estga.monument.Monument;
-import pt.estga.shared.audit.AuditedEntity;
+import pt.estga.shared.entities.AuditedEntity;
 import pt.estga.user.entities.User;
 
 import java.time.Instant;
