+++ monument/src/main/java/pt/estga/monument/Monument.java
@@ -7,7 +7,7 @@ import org.locationtech.jts.geom.GeometryFactory;
 import org.locationtech.jts.geom.Point;
 import org.locationtech.jts.geom.PrecisionModel;
 import pt.estga.territory.entities.AdministrativeDivision;
-import pt.estga.shared.audit.AuditedEntity;
+import pt.estga.shared.entities.AuditedEntity;
 
 @Entity
 @NoArgsConstructor
