+++ territory/src/main/java/pt/estga/territory/entities/AdministrativeDivision.java
@@ -3,7 +3,7 @@ package pt.estga.territory.entities;
 import jakarta.persistence.*;
 import lombok.*;
 import org.locationtech.jts.geom.Geometry;
-import pt.estga.shared.audit.AuditedEntity;
+import pt.estga.shared.entities.AuditedEntity;
 
 @Entity
 @NoArgsConstructor
